One more Windows compilation fix after r15750
BUG=None
TBR=yangguo@chromium.org

Review URL: https://codereview.chromium.org/19471006

git-svn-id: http://v8.googlecode.com/svn/branches/bleeding_edge@15754 ce2b1a6d-e550-0410-aec6-3dcde31c8c00

+++ src/circular-queue.cc
@@ -74,7 +74,7 @@ SamplingCircularQueue::SamplingCircularQueue(size_t record_size_in_bytes,
   // The distance ensures that producer and consumer never step on
   // each other's chunks and helps eviction of produced data from
   // the CPU cache (having that chunk size is bigger than the cache.)
-  const int producer_consumer_distance = (2 * chunk_size_);
+  const size_t producer_consumer_distance = (2 * chunk_size_);
   consumer_pos_->dequeue_chunk_poll_pos = buffer_ + producer_consumer_distance;
   consumer_pos_->dequeue_pos = NULL;
 }
